해시태그별 피드 모아보기 기능 추가

+++ src/main/java/com/noldaga/service/FeedService.java
@@ -34,6 +34,7 @@ public class FeedService {
 
     private final FeedRepository feedRepository;
     private final UserRepository userRepository;
+    private final StoreFeedRepository storeFeedRepository;
     private final FollowRepository followRepository;
     private final HashTagService hashTagService;
 
@@ -176,6 +177,24 @@ public class FeedService {
     }
 
     @Transactional
+    public List<FeedDto> getHashTagFeed(Long hashTagId, int page,String username){
+        //회원가입된 user인지 확인
+        User user = userRepository.findByUsername(username).orElseThrow(() ->
+                new SnsApplicationException(ErrorCode.USER_NOT_FOUND, String.format("%s not founded", username)));
+
+        Pageable pageable = PageRequest.of(page,20);
+        Page<Feed> feedListPagination = feedRepository.findAllByHashTag(hashTagId,pageable);
+
+        List<FeedDto> feedDtoList = new ArrayList<>();
+
+        feedListPagination.getContent().forEach(feed -> {
+            FeedDto feedDto = FeedDto.fromEntity(feed);
+            feedDtoList.add(feedDto);
+        });
+        return feedDtoList;
+    }
+
+    @Transactional
     public FeedDto modify(FeedModifyRequest request, Long feedId, String username) {
         //유저확인
         User user = userRepository.findByUsername(username).orElseThrow(() ->
@@ -220,10 +239,10 @@ public class FeedService {
         //해시태그 지우기
         hashTagService.deleteHashTag(feedId);
         //delete
-        feedRepository.delete(feed);
+        if(storeFeedRepository.findByFeedId(feedId)>0){//저장한 사람이 한명이라도 있으면 삭제하지않고 업데이트로 진행한다
+            feed.change("삭제된 게시물입니다.","삭제된 게시물입니다.",feed.getGroupId(),feed.getRange());
+        }
+        else
+            feedRepository.delete(feed);
     }
-
-
-
-
 }
